requirements.init.txt
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
djangorestframework==3.11.1
idna==2.10
multidict==4.7.6
Pillow==7.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pyslack==0.5.0
python-dotenv==0.14.0
pytz==2020.1
slackclient==2.8.2
sqlparse==0.3.1
yarl==1.5.1
